[ADDED] Delete Todo (User ID)

+++ ToDoApp/routers/todos.py
@@ -158,7 +158,7 @@ async def update_todo(user: user_dependency, db: db_dependency, todo_request: To
 
 # Define a route to handle DELETE requests to delete a ToDo item by its ID
 @router.delete('/todo/{todo_id}', status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
+async def delete_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
     """
     Endpoint to delete a ToDo item by its ID.
 
@@ -169,15 +169,19 @@ async def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
     Deletes the ToDo item identified by `todo_id` from the database.
     If the ToDo item with `todo_id` does not exist, raises a 404 HTTPException.
     """
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Not authenticated")
+
     # Query the database for the ToDo item with the specified ID
-    todo_model = db.query(ToDos).filter(ToDos.id == todo_id).first()
+    todo_model = db.query(ToDos).filter(ToDos.id == todo_id)\
+        .filter(ToDos.owner_id==user.get('user_id')).first()
 
     # If ToDo item with `todo_id` does not exist, raise a 404 HTTPException
     if todo_model is None:
         raise HTTPException(status_code=404, detail="ToDo not found")
 
     # Delete the ToDo item from the database
-    db.query(ToDos).filter(ToDos.id == todo_id).delete()
+    db.query(ToDos).filter(ToDos.id == todo_id).filter(ToDos.owner_id==user.get('user_id')).delete()
 
     # Commit the session to apply the deletion
     db.commit()
